Changed AiRandomSingleton to be a subclass of RandomSingleton and moved it to the ai package.
