+++ jsettlers.logic/src/jsettlers/main/JSettlersGame.java
@@ -25,6 +25,7 @@ import java.text.SimpleDateFormat;
 import java.util.Arrays;
 import java.util.Date;
 
+import jsettlers.ai.AiRandomSingleton;
 import jsettlers.ai.highlevel.AiExecutor;
 import jsettlers.common.CommonConstants;
 import jsettlers.common.map.IGraphicsGrid;
@@ -58,7 +59,6 @@ import jsettlers.logic.timer.RescheduleTimer;
 import jsettlers.network.client.OfflineNetworkConnector;
 import jsettlers.network.client.interfaces.IGameClock;
 import jsettlers.network.client.interfaces.INetworkConnector;
-import jsettlers.network.synchronic.random.AiRandomSingleton;
 import jsettlers.network.synchronic.random.RandomSingleton;
 
 /**
